pyproject.toml
[project]
name = "synth-sdk"
version = "0.2.113"
description = ""
authors = [{name = "Synth AI", email = "josh@usesynth.ai"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "pydantic",
    "requests",
    "asyncio",
    "zyk>=0.2.24",
    "build>=1.2.2.post1",
    "pypi",
    "twine>=4.0.0",
    "keyring>=24.0.0",
    "python-dotenv>=1.0.1",
    "langfuse>=2.53.9",
    "pytest>=8.3.3",
    "pydantic-openapi-schema>=1.5.1",
    "pytest-asyncio>=0.24.0",
    "boto3>=1.35.71",
    "botocore>=1.35.71",
    "tqdm>=4.66.4",
    "craftaxlm>=0.0.5",
]
classifiers = []

[project.urls]
Homepage = "https://github.com/synth-laboratories/synth-sdk"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]

[tool.hatch.build.targets.wheel]
packages = ["synth_sdk"]
